update_products_progress

+++ Poolservice/products/models.py
@@ -17,7 +17,7 @@ class Product(models.Model):
     code = models.CharField(unique=True, max_length=15 , verbose_name='Código', blank=True, null=True)
     description = models.CharField(max_length=85 , verbose_name='Descripción')
     tc = models.CharField(max_length=20 , verbose_name='Tipo de cambio' , choices=change_types, blank=True, null=True)
-    price = models.FloatField(verbose_name='Precio')
+    price = models.DecimalField(max_digits=10, decimal_places=2, verbose_name='Precio')
     stock = models.BooleanField(default=False)
 
     def __str__(self):
